Adds requirements and removes an optional
* My solution

* remove unnecessary lines

* Revert "remove unnecessary lines"

This reverts commit 31704df5e22e9701a22618597bc8ddcfa1f0eea1.

* Add example of a behavior to tack onto Item: condition_in_words

* add returns to condition_in_words

* update README with suggestioned added requirements

* update README with suggestioned added requirements

* clarify type of method

* Add refactor suggestion

* Split wave 3 into two waves

* swap_items -> swap_item; Add tests for swap_item

* Revert "update README with suggestioned added requirements"

This reverts commit bcfec2022da34de8b030a00f4ba78574876b16e3.

* add back accidentally deleted lines from README

* remove solution

* add mysteriously missing wave 6 back to README

* Add back paragraph about `condition_description`

* update optional based on our convo

* Add tests for condition_description

* fix missed swap_item naming; Add tests for condition_description

* Update README.md


* Update README.md


* Update README.md


* Remove 2nd optional about making description unique
