+++ tests/test_wave_05.py
@@ -6,126 +6,51 @@ from swap_meet.decor import Decor
 from swap_meet.electronics import Electronics
 
 
-def test_best_by_category():
-    item_a = Clothing(condition=2.0)
-    item_b = Decor(condition=2.0)
-    item_c = Clothing(condition=4.0)
-    item_d = Decor(condition=5.0)
-    item_e = Clothing(condition=3.0)
-    tai = Vendor(
-        inventory=[item_a, item_b, item_c, item_d, item_e]
-    )
-
-    best_item = tai.get_best_by_category("Clothing")
-
-    assert best_item.category == "Clothing"
-    assert best_item.condition == pytest.approx(4.0)
+def test_clothing_has_default_category_and_to_str():
+    cloth = Clothing()
+    assert cloth.category is "Clothing"
+    assert str(cloth) == "The finest clothing you could wear."
 
 
-def test_best_by_category_no_matches_is_none():
-    item_a = Decor(condition=2.0)
-    item_b = Decor(condition=2.0)
-    item_c = Decor(condition=4.0)
-    tai = Vendor(
-        inventory=[item_a, item_b, item_c]
-    )
-
-    best_item = tai.get_best_by_category("Electronics")
-
-    assert best_item is None
+def test_decor_has_default_category_and_to_str():
+    decor = Decor()
+    assert decor.category is "Decor"
+    assert str(decor) == "Something to decorate your space."
 
 
-def test_best_by_category_with_duplicates():
-    item_a = Clothing(condition=2.0)
-    item_b = Clothing(condition=4.0)
-    item_c = Clothing(condition=4.0)
-    tai = Vendor(
-        inventory=[item_a, item_b, item_c]
-    )
-
-    best_item = tai.get_best_by_category("Clothing")
-
-    assert best_item.category == "Clothing"
-    assert best_item.condition == pytest.approx(4.0)
+def test_electronics_has_default_category_and_to_str():
+    electronics = Electronics()
+    assert electronics.category is "Electronics"
+    assert str(electronics) == "A gadget full of buttons and secrets."
 
 
-def test_swap_best_by_category():
-    item_a = Decor(condition=2.0)
-    item_b = Electronics(condition=4.0)
-    item_c = Decor(condition=4.0)
-    tai = Vendor(
-        inventory=[item_a, item_b, item_c]
-    )
+def test_items_have_condition_as_float():
+    items = [
+        Clothing(condition=3.5),
+        Decor(condition=3.5),
+        Electronics(condition=3.5)
+    ]
+    for item in items:
+        assert item.condition == pytest.approx(3.5)
 
-    item_d = Clothing(condition=2.0)
-    item_e = Decor(condition=4.0)
-    item_f = Clothing(condition=4.0)
-    jesse = Vendor(
-        inventory=[item_d, item_e, item_f]
-    )
+def test_items_have_condition_descriptions_that_are_the_same_regardless_of_type():
+    items = [
+        Clothing(condition=5),
+        Decor(condition=5),
+        Electronics(condition=5)
+    ]
+    five_condition_description = items[0].condition_description()
+    assert type(five_condition_description) == str
+    for item in items:
+        assert item.condition_description() == five_condition_description
 
-    result = tai.swap_best_by_category(
-        other=jesse,
-        my_priority="Clothing",
-        their_priority="Decor"
-    )
+    items[0].condition = 1
+    one_condition_description = items[0].condition_description()
+    assert type(one_condition_description) == str
 
-    assert result is True
-    assert len(tai.inventory) is 3
-    assert len(jesse.inventory) is 3
-    assert item_c not in tai.inventory
-    assert item_f in tai.inventory
-    assert item_f not in jesse.inventory
-    assert item_c in jesse.inventory
+    for item in items:
+        item.condition = 1
+        assert item.condition_description() == one_condition_description
 
+    assert one_condition_description != five_condition_description
 
-def test_swap_best_by_category_no_match_is_false():
-    tai = Vendor(
-        inventory=[]
-    )
-
-    item_a = Clothing(condition=2.0)
-    item_b = Decor(condition=4.0)
-    item_c = Clothing(condition=4.0)
-    jesse = Vendor(
-        inventory=[item_a, item_b, item_c]
-    )
-
-    result = tai.swap_best_by_category(
-        other=jesse,
-        my_priority="Clothing",
-        their_priority="Decor"
-    )
-
-    assert result is False
-    assert len(tai.inventory) is 0
-    assert len(jesse.inventory) is 3
-    assert item_a in jesse.inventory
-    assert item_b in jesse.inventory
-    assert item_c in jesse.inventory
-
-
-def test_swap_best_by_category_no_other_match():
-    item_a = Clothing(condition=2.0)
-    item_b = Decor(condition=4.0)
-    item_c = Clothing(condition=4.0)
-    tai = Vendor(
-        inventory=[item_a, item_b, item_c]
-    )
-
-    jesse = Vendor(
-        inventory=[]
-    )
-
-    result = tai.swap_best_by_category(
-        other=jesse,
-        my_priority="Decor",
-        their_priority="Clothing"
-    )
-
-    assert result is False
-    assert len(tai.inventory) is 3
-    assert len(jesse.inventory) is 0
-    assert item_a in tai.inventory
-    assert item_b in tai.inventory
-    assert item_c in tai.inventory
